Sjg 1

+++ src/main/java/model/dto/AttendanceDTO.java
@@ -11,9 +11,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AttendanceDTO {
 
-    private int attendanceId;     // 출석 ID
-    private int studentId;        // 수강생 ID
-    private int sessionId;        // 회차 ID
+    private long attendanceId;     // 출석 ID
+    private long studentId;        // 수강생 ID
+    private long sessionId;        // 회차 ID
     private String status;         // 출석 상태 (출석/지각/결석/공결)
     private LocalDateTime checkTime; // 출석 체크 시각
     private LocalDateTime createdAt;
